pyproject.toml
[project]
name = "crimescope"
version = "0.1.0"
description = "AI-Powered Urban Safety Intelligence System"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "fastembed>=0.7.4",
    "folium>=0.20.0",
    "kaleido>=1.2.0",
    "langchain>=1.2.12",
    "lightgbm>=4.6.0",
    "litellm>=1.82.3",
    "loguru>=0.7.3",
    "marimo>=0.21.1",
    "opencv-python>=4.13.0.92",
    "pandera[polars]>=0.30.1",
    "pillow>=11.3.0",
    "plotly>=6.6.0",
    "polars>=1.39.0",
    "prophet>=1.3.0",
    "pyarrow>=23.0.1",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "ruff>=0.15.6",
    "shap>=0.51.0",
    "streamlit>=1.55.0",
    "timm>=1.0.25",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.42.0",
    "xgboost>=3.2.0",
]
